renamedComp
