+++ client/components/index.js
@@ -3,9 +3,9 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
+export {default as Navbar} from './Navbar'
+export {default as UserHome} from './UserHomePage'
+export {Login, Signup} from './AuthForm'
 export {default as AllProducts} from './ProductSubRoutes'
-export {default as ErrorPage} from './errorPage'
+export {default as ErrorPage} from './ErrorPage'
 export {default as CartView} from './CartView'
